model_training/requirements_training.txt
# Training dependencies
torch>=2.0.0
numpy>=1.24.0
datasets>=2.14.0
tqdm>=4.65.0
python-chess>=1.999

